Complete the rest of old domain

+++ src/back/Plants.Application/Requests/TotalStatsRequest.cs
@@ -5,4 +5,4 @@ namespace Plants.Application.Requests;
 public record TotalStatsRequest() : IRequest<TotalStatsResult>;
 
 public record TotalStatsResult(IEnumerable<GroupTotalStats> Groups);
-public record GroupTotalStats(int GroupId, string GroupName, decimal Income, long Instructions, long Popularity);
+public record GroupTotalStats(long GroupId, string GroupName, decimal Income, long Instructions, long Popularity);
